pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autonomous-engineer"
version = "1.0.0"
description = "Ultra Lean Autonomous Software Engineer — beats Claude Code"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "litellm>=1.40.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "llama-index>=0.10.0",
    "llama-index-embeddings-huggingface>=0.2.0",
    "qdrant-client>=1.9.0",
    "pygithub>=2.3.0",
    "gitpython>=3.1.43",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "hypothesis>=6.100.0",
]

[project.scripts]
ae = "agent.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=agent --cov=memory --cov=retrieval --cov-report=term-missing --cov-fail-under=90"
